mod-udf-enc-dec/application/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-dlp
google-api-python-client==2.5.0
google-auth==1.30.1
google-cloud-storage
google-cloud-core==1.6.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4